chore: move profile resolution
