+++ src/core/profile/index.ts
@@ -1,3 +1,4 @@
 export * from './profile';
 export * from './profile.typed';
 export * from './profile-configuration';
+export * from './resolve-profile-ast';
