+++ src/node/client/client.typed.ts
@@ -1,4 +1,9 @@
-import { NonPrimitive, TypedProfile, UsecaseType } from '../../core';
+import {
+  NonPrimitive,
+  resolveProfileAst,
+  TypedProfile,
+  UsecaseType,
+} from '../../core';
 import { NodeFileSystem } from '../filesystem';
 import { SuperfaceClientBase } from './client';
 
@@ -26,7 +31,15 @@ export function createTypedClient<TProfiles extends ProfileUseCases<any, any>>(
     public async getProfile<TProfile extends keyof TProfiles>(
       profileId: TProfile
     ): Promise<TypedProfile<TProfiles[TProfile]>> {
-      const ast = await this.internal.resolveProfileAst(profileId as string);
+      const ast = await resolveProfileAst({
+        profileId: profileId as string,
+        logger: this.logger,
+        fetchInstance: this.fetchInstance,
+        fileSystem: NodeFileSystem,
+        config: this.config,
+        crypto: this.crypto,
+        superJson: this.superJson,
+      });
       const profileConfiguration = await this.internal.getProfileConfiguration(
         ast
       );
